+++ trace/query_gtrace.go
@@ -1623,23 +1623,24 @@ func QueryOnPoolTry(t *Query, c *context.Context, call call) func(error) {
 	}
 }
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
-func QueryOnPoolWith(t *Query, c *context.Context, call call) func(_ error, attempts int) {
+func QueryOnPoolWith(t *Query, c *context.Context, call call) func(attempts int, _ error) {
 	var p QueryPoolWithStartInfo
 	p.Context = c
 	p.Call = call
 	res := t.onPoolWith(p)
-	return func(e error, attempts int) {
+	return func(attempts int, e error) {
 		var p QueryPoolWithDoneInfo
-		p.Error = e
 		p.Attempts = attempts
+		p.Error = e
 		res(p)
 	}
 }
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
-func QueryOnPoolPut(t *Query, c *context.Context, call call) func(error) {
+func QueryOnPoolPut(t *Query, c *context.Context, call call, session sessionInfo) func(error) {
 	var p QueryPoolPutStartInfo
 	p.Context = c
 	p.Call = call
+	p.Session = session
 	res := t.onPoolPut(p)
 	return func(e error) {
 		var p QueryPoolPutDoneInfo
@@ -1648,22 +1649,27 @@ func QueryOnPoolPut(t *Query, c *context.Context, call call) func(error) {
 	}
 }
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
-func QueryOnPoolGet(t *Query, c *context.Context, call call) func(error) {
+func QueryOnPoolGet(t *Query, c *context.Context, call call) func(session sessionInfo, attempts int, _ error) {
 	var p QueryPoolGetStartInfo
 	p.Context = c
 	p.Call = call
 	res := t.onPoolGet(p)
-	return func(e error) {
+	return func(session sessionInfo, attempts int, e error) {
 		var p QueryPoolGetDoneInfo
+		p.Session = session
+		p.Attempts = attempts
 		p.Error = e
 		res(p)
 	}
 }
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
-func QueryOnPoolChange(t *Query, limit int, idle int) {
+func QueryOnPoolChange(t *Query, limit int, index int, idle int, wait int, createInProgress int) {
 	var p QueryPoolChange
 	p.Limit = limit
+	p.Index = index
 	p.Idle = idle
+	p.Wait = wait
+	p.CreateInProgress = createInProgress
 	t.onPoolChange(p)
 }
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals