+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Changed `trace.Table` and `trace.Query` traces 
+
 ## v3.79.0
 * Added commit messages for topic listener
 * EOF error in RecvMsg is no longer logged