+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+## v3.79.1
 * Changed `trace.Table` and `trace.Query` traces
 * Implemented `internal/pool` the same as table client pool from `internal/table.Client`